type data

+++ inc/usart.h
@@ -26,7 +26,7 @@ static inline void usart_put(USART_t * usart, uint8_t txdata);
 static inline uint8_t usart_get(USART_t * usart);
 static inline void usart_format_set(USART_t *usart, USART_CHSIZE_t charSize,USART_PMODE_t parityMode, bool twoStopBits);
  uint8_t usart_getchar(USART_t *usart);
- void usart_putchar(USART_t *usart, uint16_t c);
+ void usart_putchar(USART_t *usart, uint8_t c);
  void usart_set_bsel_bscale_value(USART_t *usart, uint16_t bsel, uint8_t bscale);
  bool usart_set_baudrate(USART_t *usart, uint32_t baud, uint32_t cpu_hz);
